requirements.txt
streamlit
pandas
openpyxl
xlsxwriter
google-generativeai

